// app/interfaces/IColors.ts
export interface IColors {
	mainColor: string
	mainColorRegular: string
	mainColorTint: string
	mainColorTint2: string
	mainColorShade: string
	mainColorBold: string
	secondaryColorRegular: string
	bgBlack: string
	bgWhite: string
	bgGray: string
	frBlack: string
	frWhite: string
	frTextGrayPrimary: string
	frTextGraySecondary: string
	frGreenTint2: string
	bgInput: string
	fgInput: string
	frOrange: string
	strokeIconPrimary: string
	strokeIconTint: string
	strokeIconShade: string
	secondaryRed: string
	secondaryGreen: string
	secondaryGreenTint1: string
	secondaryGreenTint2: string
	secondaryGreenShade1: string
	secondaryGreenShade2: string
}